# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API documentation using Swagger",
        "contact": {
            "email": "linsmar_cruz@hotmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/scheduleTypes": {
            "get": {
                "tags": [
                    "ScheduleTypes"
                ],
                "summary": "Display a listing of the scheduleTypes",
                "description": "Get all scheduleTypes",
                "operationId": "57ee339685138514f48a4677b136d909",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ScheduleType"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/scheduleTypes/{uuid}": {
            "get": {
                "tags": [
                    "ScheduleTypes"
                ],
                "summary": "Display the specified scheduleType",
                "description": "Get a scheduleType by UUID",
                "operationId": "77e2b84f6d906f87936b3515a0a64f2f",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ScheduleType"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "ScheduleType not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "ScheduleTypes"
                ],
                "summary": "Update the specified scheduleType in storage",
                "description": "Update a scheduleType by UUID",
                "operationId": "0d607609d0cfcd2bfb4cccf128a54c1a",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ScheduleType"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ScheduleType updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ScheduleType"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "ScheduleType not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "ScheduleTypes"
                ],
                "summary": "Remove the specified scheduleType from storage",
                "description": "Delete a scheduleType by UUID",
                "operationId": "df171c0ff7c0225b9edba9a67069167f",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "ScheduleType deleted successfully"
                    },
                    "404": {
                        "description": "ScheduleType not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/scheduleTypes/": {
            "post": {
                "tags": [
                    "ScheduleTypes"
                ],
                "summary": "Store a new scheduleType in storage",
                "description": "Create a new scheduleType",
                "operationId": "23890e6f596283ce0800cae40d340cf0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ScheduleType"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "ScheduleType created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ScheduleType"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ScheduleType": {
                "title": "ScheduleType",
                "properties": {
                    "uuid": {
                        "type": "integer"
                    },
                    "description": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}